Moved Loading component to global
