+++ src/Core/components/SignUpForm/subs/2_SignupForms/1_InitialSignUp.js
@@ -10,7 +10,7 @@ import Icon from '../../../../../global/icons/Icon';
 import { ICONS } from '../../../../../global/icons/iconConstants';
 
 // Components
-import Loading from '../../../Loading';
+import Loading from '../../../../../global/components/Loading';
 
 const CHECK_EMAIL = gql`
 	mutation checkEmail($email: String!) {
