+++ src/global/components/Loading/index.js
